requirements.txt
Django==2.2.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
